添加版本同步消息

+++ src/main/java/com/tct/mapper/DeviceGunMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface DeviceGunMapper {
-    long countByExample(DeviceGunExample example);
+    int countByExample(DeviceGunExample example);
 
     int deleteByExample(DeviceGunExample example);
